Refactor project name from 's_e_c' to 'sec' across all files and configurations
