+++ README.md
@@ -1,4 +1,4 @@
-# s_e_c
+# sec
 
 A new Flutter project.
