+++ macos/Runner/Configs/AppInfo.xcconfig
@@ -5,10 +5,10 @@
 // 'flutter create' template.
 
 // The application's name. By default this is also the title of the Flutter window.
-PRODUCT_NAME = s_e_c
+PRODUCT_NAME = sec
 
 // The application's bundle identifier
-PRODUCT_BUNDLE_IDENTIFIER = com.example.sEc
+PRODUCT_BUNDLE_IDENTIFIER = com.example.sec
 
 // The copyright displayed in application information
 PRODUCT_COPYRIGHT = Copyright © 2025 com.example. All rights reserved.
