+++ web/manifest.json
@@ -1,6 +1,6 @@
 {
-    "name": "s_e_c",
-    "short_name": "s_e_c",
+    "name": "sec",
+    "short_name": "sec",
     "start_url": ".",
     "display": "standalone",
     "background_color": "#0175C2",
