+++ windows/runner/Runner.rc
@@ -90,12 +90,12 @@ BEGIN
         BLOCK "040904e4"
         BEGIN
             VALUE "CompanyName", "com.example" "\0"
-            VALUE "FileDescription", "s_e_c" "\0"
+            VALUE "FileDescription", "sec" "\0"
             VALUE "FileVersion", VERSION_AS_STRING "\0"
-            VALUE "InternalName", "s_e_c" "\0"
+            VALUE "InternalName", "sec" "\0"
             VALUE "LegalCopyright", "Copyright (C) 2025 com.example. All rights reserved." "\0"
-            VALUE "OriginalFilename", "s_e_c.exe" "\0"
-            VALUE "ProductName", "s_e_c" "\0"
+            VALUE "OriginalFilename", "sec.exe" "\0"
+            VALUE "ProductName", "sec" "\0"
             VALUE "ProductVersion", VERSION_AS_STRING "\0"
         END
     END
